+++ fastly/ngwaf/v1/timeseries/doc.go
@@ -1,0 +1,3 @@
+// Package timeseries contains API operations to get
+// Fastly Next-Gen WAF time series
+package timeseries