+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Enhancements:
 - feat(ngwaf): add support for requests ([#688](https://github.com/fastly/go-fastly/pull/688))
+- feat(ngwaf): add support for timeseries ([#689](https://github.com/fastly/go-fastly/pull/689))
 
 ### Bug fixes: